// src/app/interfaces/cancion.ts
export interface Cancion {
    acousticness: string
    album_name: string
    album_release_date: string
    artists: string
    country: string
    daily_movement: string
    daily_rank: string
    danceability: string
    duration_ms: string
    energy: string
    instrumentalness: string
    is_explicit: string
    key: string
    liveness: string
    loudness: string
    mode: string
    name: string
    popularity: string
    snapshot_date: string
    speechiness: string
    spotify_id: string
    tempo: string
    time_signature: string
    valence: string
    weekly_movement: string
}